+++ tests/test_graph_utils.py
@@ -4,29 +4,29 @@ import pydot
 import os
 import networkx as nx
 from pyformlang.finite_automaton import State, Symbol, NondeterministicFiniteAutomaton
-from project import work_with_graph
+from project import graph_utils
 
 
 def test_get_nodes_edges_labels_created_graph():
     graph = cfpq_data.labeled_cycle_graph(4, "a")
 
-    assert work_with_graph.get_nodes_edges_labels(graph) == (4, 4, {"a"})
+    assert graph_utils.get_nodes_edges_labels(graph) == (4, 4, {"a"})
 
 
 def test_get_nodes_edges_labels_downloaded_graph():
     bzip_path = cfpq_data.download("bzip")
     bzip = cfpq_data.graph_from_csv(bzip_path)
 
-    assert work_with_graph.get_nodes_edges_labels(bzip)[0] == 632
-    assert work_with_graph.get_nodes_edges_labels(bzip)[1] == 556
+    assert graph_utils.get_nodes_edges_labels(bzip)[0] == 632
+    assert graph_utils.get_nodes_edges_labels(bzip)[1] == 556
     assert True == (
-        work_with_graph.get_nodes_edges_labels(bzip)[2] == {"a", "d"}
-        or work_with_graph.get_nodes_edges_labels(bzip)[2] == {"d", "a"}
+        graph_utils.get_nodes_edges_labels(bzip)[2] == {"a", "d"}
+        or graph_utils.get_nodes_edges_labels(bzip)[2] == {"d", "a"}
     )
 
 
 def test_save_two_cycles_graph_file_existing():
-    work_with_graph.save_two_cycles_graph(
+    graph_utils.save_two_cycles_graph(
         3, 4, ("a", "b"), "save_two_cycles_graph_output.dot"
     )
 
@@ -38,7 +38,7 @@ def test_save_two_cycles_graph_file_existing():
 def test_save_two_cycles_graph_save_correction():
     graph = cfpq_data.labeled_two_cycles_graph(3, 4, labels=("a", "b"))
 
-    work_with_graph.save_two_cycles_graph(
+    graph_utils.save_two_cycles_graph(
         3, 4, ("a", "b"), "save_two_cycles_graph_output.dot"
     )
     saved_graph = pydot.graph_from_dot_file("save_two_cycles_graph_output.dot")
@@ -55,13 +55,13 @@ def test_save_two_cycles_graph_save_correction():
 
 
 def test_create_dfa_by_regex_determinism():
-    dfa = work_with_graph.create_dfa_by_regex("abc|d")
+    dfa = graph_utils.create_dfa_by_regex("abc|d")
 
     assert dfa.is_deterministic()
 
 
 def test_create_dfa_by_regex_acceptance():
-    dfa = work_with_graph.create_dfa_by_regex("abc|d")
+    dfa = graph_utils.create_dfa_by_regex("abc|d")
 
     assert dfa.accepts(["abc"])
 
@@ -70,7 +70,7 @@ def test_create_nfa_by_graph_without_start_final_nodes_created_by_download():
     bzip_path = cfpq_data.download("bzip")
     bzip = cfpq_data.graph_from_csv(bzip_path)
 
-    nfa = work_with_graph.create_nfa_by_graph(bzip)
+    nfa = graph_utils.create_nfa_by_graph(bzip)
 
     assert nfa.accepts(["d"])
     assert nfa.accepts(["a", "a"])
@@ -80,7 +80,7 @@ def test_create_nfa_by_graph_with_start_final_nodes_created_by_download():
     bzip_path = cfpq_data.download("bzip")
     bzip = cfpq_data.graph_from_csv(bzip_path)
 
-    nfa = work_with_graph.create_nfa_by_graph(bzip, [223, 102], [257, 422])
+    nfa = graph_utils.create_nfa_by_graph(bzip, [223, 102], [257, 422])
 
     assert nfa.accepts(["d"])
     assert nfa.accepts(["a", "a"])
@@ -90,7 +90,7 @@ def test_create_nfa_by_graph_with_start_final_nodes_created_by_download():
 def test_create_nfa_by_graph_without_start_final_nodes_created_by_function():
     graph = cfpq_data.labeled_two_cycles_graph(3, 4, labels=("a", "b"))
 
-    nfa = work_with_graph.create_nfa_by_graph(graph)
+    nfa = graph_utils.create_nfa_by_graph(graph)
 
     assert nfa.accepts(["a"])
     assert nfa.accepts(["b", "a"])
@@ -100,44 +100,8 @@ def test_create_nfa_by_graph_without_start_final_nodes_created_by_function():
 def test_create_nfa_by_graph_with_start_final_nodes_created_by_function():
     graph = cfpq_data.labeled_two_cycles_graph(3, 4, labels=("a", "b"))
 
-    nfa = work_with_graph.create_nfa_by_graph(graph, [6], [1])
+    nfa = graph_utils.create_nfa_by_graph(graph, [6], [1])
 
     assert not nfa.accepts(["b", "a"])
     assert nfa.accepts(["b", "b", "a"])
     assert not nfa.accepts(["a", "b"])
-
-
-def test_make_regular_path_query_with_empty_graph():
-    graph = nx.MultiDiGraph()
-    result = work_with_graph.make_regular_path_query("", graph)
-
-    assert result == set()
-
-
-def test_make_regular_path_query_with_simple_graph_and_nonempty_intersection():
-    automaton = NondeterministicFiniteAutomaton()
-    automaton.add_transition(State(0), Symbol("a"), State(1))
-    automaton.add_transition(State(1), Symbol("a"), State(0))
-    graph = automaton.to_networkx()
-    result = work_with_graph.make_regular_path_query("a|b", graph)
-
-    assert result == {(0, 1), (1, 0)}
-
-
-def test_make_regular_path_query_with_simple_graph_and_empty_intersection():
-    automaton = NondeterministicFiniteAutomaton()
-    automaton.add_transition(State(0), Symbol("a"), State(1))
-    automaton.add_transition(State(1), Symbol("a"), State(0))
-    graph = automaton.to_networkx()
-    result = work_with_graph.make_regular_path_query("b*", graph)
-
-    assert result == set()
-
-
-def test_make_regular_path_query_labeled_two_cycles_graph():
-    graph = cfpq_data.labeled_two_cycles_graph(2, 2, labels=("a", "b"))
-    result = work_with_graph.make_regular_path_query(
-        "(a + b)* b (a + b)*", graph, [0, 1], [2, 3]
-    )
-
-    assert result == {(1, 2), (0, 2), (0, 3), (1, 3)}
